pyproject.toml
[tool.poetry]
# $ pip install pytemplate
name = "classifiers"
version = "0.0.1"
description = "Popular classifier implementations which can be used for pretraining backbones for other downstream tasks."
authors = ["Bradley Selee"]
license = "LICENSE.txt"

[tool.isort]
profile = "black" # prevents isort from reformatting black

# This field lists other packages that your project depends on to run.
# Any package you put here will be installed by pip when your project is
# installed, so they must be valid existing projects.
[tool.poetry.dependencies]
python = ">3.9"
matplotlib = "*"
pandas = "*"
pytest = "*"
pyyaml = "*"
fire = "*"
torch = "*"
torchvision = "*"
tqdm = "*"
pycocotools = "*"

# Linters/formatters
mypy = "*"
pylint = "*"
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"